requirements.txt
neo4j-driver
pylyrics3
selenium
flask


